requirements.txt
constants==0.6.0
dnspython==2.3.0
Quart==0.19.6
pymongo==4.4.1
pyromod==2.0.0
Markdown==3.4.3
pyrofork
TgCrypto==1.2.5
PyYAML==6.0
tldextract==3.4.4
requests-toolbelt==1.0.0
mail-parser==3.15.0
regex
